display lehrveranstaltungs in modul list and modul details
Automatic modul-lehrveranstaltung association does not seem to work when
modul_id field is a string. Maybe we need to change the primary key of modul
to its modulnummer.
Right now we are explicitly accessing the lehrveranstaltungs table like this:
Lehrveranstaltung.where(modul_id: modul.nummer).

Remove lehrveranstaltungs reference from modul table, because we store the
references in the lehrveranstaltungs table.
Adjust example seed data to the new schema.